==> 1. IPTC news categorisation/3 Final/code/notebooks/00_taxonomy_processing.ipynb <==
import sys
sys.path.append('../')

import pandas as pd
import os
from src.embeddings import OpenaiAdaEmbedding, BertEmbedding, AngleEmbedding, MiniLMEmbedding

# Dataset preparation

First we read taxonomy from original IPTC taxonomy file and process it 

df = pd.read_excel('../data/taxonomy/TAKSONOMIJA.xlsx', sheet_name='SubjectCodes', usecols='A:E', nrows=938, skiprows=2)

Based on QCodes of each IPTC category we extract their hierarchies

def get_hierarchy(x):
    n = x.split(':')[1]
    if n[2:] == '0' * len(n[2:]):
        return 1
    elif n[5:] == '0' * len(n[5:]):
        return 2
    else:
        return 3

df['hierarchy'] = df['Qcode'].apply(lambda x: get_hierarchy(x))

df = df[['Qcode', 'hierarchy', 'IPTC NAME', 'IPTC DESCRIPTION']]

df.columns = ['code', 'hierarchy', 'name', 'description']

df

# Embedding extraction

First we instantiate classes responsible for gathering:
1. Ada Embeddings from OpenAI API
2. Embeddings from open-source, state-of-the-art `UAE-Large-V1` model https://huggingface.co/WhereIsAI/UAE-Large-V1
3. Embeddings from open-source `all-MiniLM-L6-v2` model (used inside vector database later on)  https://huggingface.co/sentence-transformers/all-MiniLM-L6-v2

To generate Ada Embeddings, an openai_key is needed. One can create such key at https://platform.openai.com/api-keys

api_key = os.environ['OPENAI_API_KEY']
openai_ada = OpenaiAdaEmbedding(api_key = api_key)
bert = BertEmbedding(model_name = 'bert-base-uncased')
angle = AngleEmbedding()
mini_lm = MiniLMEmbedding()

We extract Ada, BERT and AnglE embeddings from:
1. Names of categories
2. Descriptions of categories, as they might be more insightful

df_openai_ada = pd.DataFrame({'code': df['code'], 
                              'hierarchy': df['hierarchy'],
                              'name': df['name'], 
                              'description': df['description'], 
                              'embedding_name': df['name'].apply(lambda x: openai_ada.get_embedding(x)).apply(lambda x: x.tolist()),
                              'embedding_description': df['description'].apply(lambda x: openai_ada.get_embedding(x)).apply(lambda x: x.tolist())})

df_angle = pd.DataFrame({'code': df['code'],
                         'hierarchy': df['hierarchy'],
                         'name': df['name'],
                         'description': df['description'],
                         'embedding_name': df['name'].apply(lambda x: angle.get_embedding(x)).apply(lambda x: x.tolist()),
                         'embedding_description': df['description'].apply(lambda x: angle.get_embedding(x)).apply(lambda x: x.tolist())})

df_mini_lm = pd.DataFrame({'code': df['code'],
                           'hierarchy': df['hierarchy'],
                           'name': df['name'], 
                           'description': df['description'],
                           'embedding_name': df['name'].apply(lambda x: mini_lm.get_embedding(x)),
                           'embedding_description': df['description'].apply(lambda x: mini_lm.get_embedding(x))})

Finally we save the processed taxonomy and the dataframes containing embeddings in csv format.

df.to_csv('../data/taxonomy/taxonomy.csv', index=False)

df_openai_ada.to_csv('../data/taxonomy/embeddings/taxonomy_openai_ada.csv', index=False)
df_angle.to_csv('../data/taxonomy/embeddings/taxonomy_angle.csv', index=False)
df_mini_lm.to_csv('../data/taxonomy/embeddings/taxonomy_mini_lm.csv', index=False)

df.head(5)
