==> 1. IPTC news categorisation/3 Final/code/notebooks/08_add_articles_to_vector_database.ipynb <==
import autorootcwd
import chromadb
import requests
import os
import pandas as pd
from chromadb.config import Settings
from chromadb.utils import embedding_functions
from src.exploratory_data_analysis import gather_data
from sentence_transformers import SentenceTransformer
import uuid

articles = gather_data(path='data/2023_articles_en/')

articles = articles[~articles['categories'].str.contains('AD|SE|RU', case=False, na=False)]

articles

# Pre-processing of article texts

## Remove schedules and reviews

articles = articles[~articles['categories'].str.contains('AD|SE|RU', case=False, na=False)]
articles = articles[~articles['keywords'].str.contains('REVIEW', case=False, na=False)]
articles_df = pd.DataFrame(articles[['headline', 'lede', 'text']])
articles_df.dropna(inplace=True)

articles_df

## Remove HTML tags, split per section

import re
from html import unescape

def remove_html_tags(text):
    clean = re.compile('<.*?>', re.DOTALL)
    return re.sub(clean, '', text)

def split_into_sections(text):
    return re.split(r'\n\n+', text)

articles_df['text_processed'] = articles_df['text'].apply(unescape)
articles_df['text_processed'] = articles_df['text_processed'].apply(remove_html_tags)
articles_df['text_processed'] = articles_df['text_processed'].apply(split_into_sections)


# Embedding of headlines, lede and texts (around 20min of computations)

model = SentenceTransformer("sentence-transformers/all-MiniLM-L6-v2")

def get_embeddings(text):
    return list(model.encode(text))


articles_df['headline_embeddings'] = articles_df['headline'].apply(get_embeddings)
articles_df['lede_embeddings'] = articles_df['lede'].apply(get_embeddings)
articles_df['text_mean_embeddings'] = articles_df['text_embeddings'].apply(lambda embeddings: [sum(x) / len(x) for x in zip(*embeddings)])

# Averaging the headline, lede and text embeddings, to get the article embedding

articles_df['average_embeddings'] = articles_df.apply(lambda row: [(a + b + c) / 3 for a, b, c in zip(row['text_mean_embeddings'], row['headline_embeddings'], row['lede_embeddings'])], axis=1)
articles_df['full_text'] = 'Headline: ' + articles_df['headline'] + '\nLede: ' + articles_df['lede'] + '\nText: ' + articles_df['text']

# Save result to .csv

articles_chroma = pd.DataFrame(articles_df[['full_text', 'average_embeddings']])
articles_chroma.columns = ['text', 'embeddings']
articles_chroma.to_csv('application/articles_chroma.csv', index=False)

# Adding most similar articles (in terms of cosine similarity) to IPTC categories to ChromaDB.

client = chromadb.HttpClient(host='localhost', port=8000)
collection = client.get_collection("database")

embeddings = list(articles_chroma['embeddings'])
texts = list(articles_chroma['text'])

query_result = collection.query(query_embeddings=embeddings, n_results=1)

articles = []
metadatas = []
embeddings2 = []

for i in range(len(query_result['ids'])):
    if 0.5 < query_result['distances'][i][0] < 0.9:
        articles.append(texts[i])
        metadatas.append(query_result['metadatas'][i][0])
        embeddings2.append(embeddings[i])

ids = [str(uuid.uuid4()) for i in range(len(articles))]
collection.add(
    ids =  ids,
    documents = articles,
    metadatas = metadatas,
    embeddings = embeddings2
    )
